refactor: move statements and parsers to proper files
This way we can divide between the AST logic and the parser logic, and find it with more ease
