--- src/ansi/data_type_structures.rs
@@ -1,2 +1,0 @@
-pub mod ast;
-pub mod parser;
